+++ dev/src/pattern_lexer.js
@@ -104,7 +104,7 @@
             return captureVals(TOKENS.OP.rgx, pattern);
         }
 
-        function compilePattern(pattern) {
+        function compilePattern(pattern, ignoreCase) {
             pattern = pattern || '';
 
             if(pattern){
@@ -131,7 +131,7 @@
                 //single slash is treated as empty and end slash is optional
                 pattern += '\\/?';
             }
-            return new RegExp('^'+ pattern + '$', 'i');
+            return new RegExp('^'+ pattern + '$', ignoreCase? 'i' : '');
         }
 
         function replaceTokens(pattern, regexpName, replaceName) {