+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Crossroads.js Changelog #
 
+## ignorecase branch
+
+ - changed `Route.rules` array validation to be case insensitive (#49)
+ - route matching is case insensitive. (#53)
+ - added way to toggle case sensitivity `Route.ignoreCase` (#53)
+
 
 ## Next ##
 
@@ -18,8 +24,6 @@
  - query string support affected old segment rules, now `?` is considered as
    a segment divisor as `/` otherwise optional query string RegExp wouldn't
    match proper segment if following a required segment. (#33)
- - changed `Route.rules` array validation to be case insensitive (#49)
- - route matching is case insensitive. (#53)